feat: created a AfterLogin page
Now when the user logs in or signs up it takes us directly to the AfterLogin page which right now just has a navbar with a logout button. (The button works)
And I changed the name of the homepage that we have before to BeforeLogin and put this to components (after and before login) to a homepages component.