fix: fixed a bug
The bug was that when we refreshed the page, we were logged out, even though we were just logged in.
I saved the user information in localstorage and now when we refresh the page, first it checks that if we have user information or not and if we have it, we're automaticlly signed in and the after login page will apper.