feat: added a LogoutRedirect component
This component is for the times that we're already logged in but we go to the signup page. (through the url)
So now when we're logged in, if we go to the signup or login page we see this component which says to us that we're logged in already and if we want to create another account or login with another account we need to logout from current account first.
And there is a button too. For logout.